+++ tests/pjit_test.py
@@ -1084,7 +1084,7 @@ class PJitTest(jtu.BufferDonationTestCase):
       with self.assertRaisesRegex(
           ValueError,
           r"One of with_sharding_constraint.*Sharding "
-          r"MeshPspecSharding\(mesh={'replica': 1, 'data': 1, 'mdl': 2}, "
+          r"NamedSharding\(mesh={'replica': 1, 'data': 1, 'mdl': 2}, "
           r"partition_spec=PartitionSpec\(None, \('mdl',\), None, None\)\) is only "
           "valid for values of rank at least 4, but was applied to a value of rank 1"):
         pjit_f(jnp.array([1, 2, 3]))